+++ src/com/mikedeejay2/voxel/engine/graphics/shaders/StaticShader.java
@@ -2,6 +2,7 @@ package com.mikedeejay2.voxel.engine.graphics.shaders;
 
 import com.mikedeejay2.voxel.engine.graphics.objects.Camera;
 import com.mikedeejay2.voxel.engine.utils.Maths;
+import org.joml.Matrix4d;
 import org.joml.Matrix4f;
 import org.joml.Vector3f;
 
@@ -42,7 +43,7 @@ public class StaticShader extends ShaderProgram
         super.loadVector(location_skyColor, new Vector3f(r, g, b));
     }
 
-    public void loadTransformationMatrix(Matrix4f matrix)
+    public void loadTransformationMatrix(Matrix4d matrix)
     {
         super.loadMatrix(location_transformationMatrix, matrix);
     }
@@ -53,7 +54,7 @@ public class StaticShader extends ShaderProgram
         super.loadMatrix(location_viewMatrix, camera.getViewMatrix());
     }
 
-    public void loadProjectionMatrix(Matrix4f projection)
+    public void loadProjectionMatrix(Matrix4d projection)
     {
         super.loadMatrix(location_projectionMatrix, projection);
     }
